Move hooks functions to src/hooks
