+++ src/components/Button/Button.tsx
@@ -1,5 +1,5 @@
 import React, { ComponentProps, forwardRef } from "react";
-import { themable, ThemedComponent, ThemedComponentProps } from "../../util/theme";
+import { themable, ThemedComponent, ThemedComponentProps } from "../../hooks/theme";
 import cssClasses from "./Button.module.scss";
 
 export const Button = themable<ComponentProps<"button">>(
