+++ src/components/ElectionResultsTableSection/ElectionResultsTableSection.tsx
@@ -3,7 +3,7 @@ import { ResultsTable } from "../ResultsTable/ResultsTable";
 import { Heading2 } from "../Typography/Typography";
 import { electionHasSeats, ElectionBallotMeta, ElectionResults, ElectionResultsCandidate } from "../../types/Election";
 import { formatGroupedNumber, formatPercentage, fractionOf } from "../../util/format";
-import { ClassNames, themable } from "../../util/theme";
+import { ClassNames, themable } from "../../hooks/theme";
 import cssClasses from "./ElectionResultsTableSection.module.scss";
 import { Button } from "../Button/Button";
