+++ src/components/Typography/Typography.tsx
@@ -1,5 +1,5 @@
 import React, { ComponentProps, ComponentType, forwardRef } from "react";
-import { mergeClasses, PropsObject, themable, ThemableComponent, ThemedComponentProps } from "../../util/theme";
+import { mergeClasses, PropsObject, themable, ThemableComponent, ThemedComponentProps } from "../../hooks/theme";
 import cssClasses from "./Typography.module.scss";
 
 export function makeTypographyComponent<Props extends PropsObject>(
