+++ src/index.ts
@@ -1,8 +1,12 @@
-export * from "./util/theme";
 export * from "./util/api";
 export * from "./util/electionApi";
+
 export * from "./constants/servers";
+
 export * from "./types/Election";
+
+export * from "./hooks/theme";
+export * from "./hooks/useApiResponse";
 
 export * from "./components/Typography/Typography";
 export * from "./components/Button/Button";
